createMapMarker refactored
